[FIX] point_of_sale: Changing the report.pos.order.product_qty field type to support negative quantities from refund Orders

+++ addons/point_of_sale/report/pos_order_report.py
@@ -26,7 +26,7 @@ class PosOrderReport(models.Model):
     average_price = fields.Float(string='Average Price', readonly=True, group_operator="avg")
     company_id = fields.Many2one('res.company', string='Company', readonly=True)
     nbr_lines = fields.Integer(string='Sale Line Count', readonly=True)
-    product_qty = fields.Integer(string='Product Quantity', readonly=True)
+    product_qty = fields.Float(string='Product Quantity', readonly=True)
     journal_id = fields.Many2one('account.journal', string='Journal')
     delay_validation = fields.Integer(string='Delay Validation')
     product_categ_id = fields.Many2one('product.category', string='Product Category', readonly=True)
